+++ src/main/java/Controller/LogoutController.java
@@ -2,6 +2,7 @@ package Controller;
 
 import java.io.IOException;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -37,6 +38,8 @@ public class LogoutController extends HttpServlet {
 	protected void requestPro(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 		request.getSession().invalidate();
-		response.sendRedirect(request.getContextPath() + "/view/login/login.jsp");
+//		response.sendRedirect(request.getContextPath() + "/view/login/login.jsp");
+		RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/view/login/login.jsp");
+		dispatcher.forward(request, response);
 	}
 }
