+++ src/main/webapp/WEB-INF/components/aside.jsp
@@ -1,0 +1,24 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+<link rel="stylesheet" href="${pageContext.request.contextPath}/css/components/aside.css?ver=1">
+</head>
+<body>
+	<div class="asideContainer">
+		<div class="asideTitleContainer">
+			<img/>
+			<p>메인 메뉴</p>
+		</div>
+		<ul class="asideMenuContainer">
+			<li class="asideMenu"><a href="">서브메뉴1</a></li>
+			<li class="asideMenu"><a href="">서브메뉴2</a></li>
+			<li class="asideMenu"><a href="">서브메뉴3</a></li>
+			<li class="asideMenu"><a href="">서브메뉴4</a></li>
+		</ul>
+	</div>
+</body>
+</html>
