+++ src/main/webapp/WEB-INF/components/inputTable/inputTable.js
