+++ src/main/webapp/WEB-INF/view/loan/productRegistration/productRegistration.jsp
