+++ src/main/webapp/WEB-INF/view/loan/productSubscription/productSubscription.js
