+++ src/main/webapp/WEB-INF/view/login/tempRegister.jsp
