+++ src/main/webapp/WEB-INF/web.xml
@@ -13,6 +13,16 @@
 		<welcome-file>default.jsp</welcome-file>
 	</welcome-file-list>
 	<session-config>
-		<session-timeout>10</session-timeout>
+		<session-timeout>100</session-timeout>
 	</session-config>
+	<servlet>
+		<servlet-name>MyServlet</servlet-name>
+		<servlet-class>Controller.NavigationController</servlet-class>
+	</servlet>
+	
+	<servlet-mapping>
+		<servlet-name>MyServlet</servlet-name>
+		<url-pattern>/navigation/*</url-pattern>
+	</servlet-mapping>
+	
 </web-app>
