--- src/main/webapp/components/header/header.css
@@ -1,48 +1,0 @@
-@charset "UTF-8";
-
-header {
-	width: 100%;
-	height: 80px;
-	position: fixed;
-	top: 0;
-	z-index: 10;
-	display: flex;
-	background-color: #ffffff;
-	border-bottom: 1px solid #2DB400;
-	box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.25);
-}
-.headerLogoContainer {
-	width: 260px;
-	height: 100%;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-}
-.headerLogo {
-	cursor: pointer;
-	font-size: 25px;
-	font-weight: bold;
-	line-height: 100%;
-}
-.headerMenuContainer {
-	width: 400px;
-	height: 100%;
-	display: flex;
-	justify-content: space-around;
-	align-items: center;
-	margin: 0;
-	padding: 0;
-}
-.headerMenuContainer > li {
-	cursor: pointer;
-	list-style: none;
-}
-.headerMenuContainer > li > a {
-	font-size: 18px;
-	font-weight: bold;
-}
-.headerMenuContainer > li > a:hover {
-	color: var(--primary-color);
-	font-weight: bold;
-	text-decoration: none;
-}
