--- src/main/webapp/components/header/userInformation.css
@@ -1,44 +1,0 @@
-@charset "UTF-8";
-
-.headerSessionInformationContainer {
-	width: fit-content;
-	height: 100%;
-	gap: 30px;
-	display: flex;
-	justify-content: flex-start;
-	align-items: center;
-	padding: 0px 72px;
-	box-sizing: border-box;
-	right: 0;
-	position: absolute;
-}
-.headerSessionTimeout {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	gap: 10px;
-}
-.headerSessionButton {
-	width: 50px;
-	height: 24px;
-	border: 1px solid #2EB400;
-	border-radius: 5px;
-	color: #2EB400;
-	display: flex;
-	align-items: center;
-	justify-content: center;	
-	cursor: pointer;
-	background: transparent;
-}
-.headerSessionButton:hover {
-	opacity: 0.8;
-	background: buttonface;
-}
-.headerSessionName {
-	font-weight: bold;
-}
-.headerSessionLogoutButton{
-	background: transparent;
-	cursor: pointer;
-	border: none;
-}
