+++ src/main/webapp/css/components/header.css
@@ -1,0 +1,90 @@
+@charset "UTF-8";
+
+header {
+	width: 100%;
+	height: 80px;
+	position: fixed;
+	top: 0;
+	z-index: 10;
+	display: flex;
+	background-color: #ffffff;
+	border-bottom: 1px solid #2DB400;
+	box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.25);
+}
+.headerLogoContainer {
+	width: 260px;
+	height: 100%;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
+.headerLogo {
+	cursor: pointer;
+	font-size: 25px;
+	font-weight: bold;
+	line-height: 100%;
+}
+.headerMenuContainer {
+	width: 400px;
+	height: 100%;
+	display: flex;
+	justify-content: space-around;
+	align-items: center;
+	margin: 0;
+	padding: 0;
+}
+.headerMenuContainer > li {
+	cursor: pointer;
+	list-style: none;
+}
+.headerMenuContainer > li > a {
+	font-size: 18px;
+	font-weight: bold;
+}
+.headerMenuContainer > li > a:hover {
+	color: var(--primary-color);
+	font-weight: bold;
+	text-decoration: none;
+}
+.headerSessionInformationContainer {
+	width: fit-content;
+	height: 100%;
+	gap: 30px;
+	display: flex;
+	justify-content: flex-start;
+	align-items: center;
+	padding: 0px 72px;
+	box-sizing: border-box;
+	right: 0;
+	position: absolute;
+}
+.headerSessionTimeout {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	gap: 10px;
+}
+.headerSessionButton {
+	width: 50px;
+	height: 24px;
+	border: 1px solid #2EB400;
+	border-radius: 5px;
+	color: #2EB400;
+	display: flex;
+	align-items: center;
+	justify-content: center;	
+	cursor: pointer;
+	background: transparent;
+}
+.headerSessionButton:hover {
+	opacity: 0.8;
+	background: buttonface;
+}
+.headerSessionName {
+	font-weight: bold;
+}
+.headerSessionLogoutButton{
+	background: transparent;
+	cursor: pointer;
+	border: none;
+}
