+++ src/main/webapp/css/components/searchResultTable.css
