+++ src/main/webapp/js/components/searchLayout.js
